+++ src/ninja/mbedded/ninjaterm/model/terminal/txRx/formatting/Formatting.java
@@ -11,11 +11,24 @@ import javafx.beans.property.SimpleObjectProperty;
  */
 public class Formatting {
 
+    //================================================================================================//
+    //========================================= CLASS ENUMS ==========================================//
+    //================================================================================================//
+
     public enum EnterKeyBehaviour {
         CARRIAGE_RETURN,
         NEW_LINE,
         CARRIAGE_RETURN_AND_NEW_LINE,
     }
+
+    public enum TxCharSendingOptions {
+        SEND_TX_CHARS_IMMEDIATELY,
+        SEND_TX_CHARS_ON_ENTER
+    }
+
+    //================================================================================================//
+    //========================================= CLASS FIELDS =========================================//
+    //================================================================================================//
 
     /**
      * Controls what is send to the COM port (TX) when "ENTER" is pressed.
@@ -23,5 +36,7 @@ public class Formatting {
      * Sending a new line is the most common behaviour.
      */
     public SimpleObjectProperty<EnterKeyBehaviour> selEnterKeyBehaviour = new SimpleObjectProperty<>(EnterKeyBehaviour.NEW_LINE);
+
+    public SimpleObjectProperty<TxCharSendingOptions> selTxCharSendingOption = new SimpleObjectProperty<>(TxCharSendingOptions.SEND_TX_CHARS_IMMEDIATELY);
 
 }